fix(tests): fix template loading in test files
avoid requirejs in such cases so there is no extra prefix
for template files
