+++ client/app/index.html
@@ -39,6 +39,8 @@
     <script src="scripts/superdesk/editor/editor.js"></script>
     <script src="scripts/superdesk/features/features.js"></script>
     <script src="scripts/superdesk/services/asset.js"></script>
+    <script src="scripts/superdesk/list/list.js"></script>
+
     <script src="scripts/superdesk/privileges/privileges.js"></script>
     <script src="scripts/superdesk/notification/notification.js"></script>
     <script src="scripts/superdesk/itemList/itemList.js"></script>
@@ -73,6 +75,8 @@
     <script src="scripts/superdesk-stream/stream.js"></script>
     <script src="scripts/superdesk-packaging/packaging.js"></script>
     <script src="scripts/superdesk-highlights/highlights.js"></script>
+
+    <script src="scripts/superdesk-archive/directives.js"></script>
     <!-- endbuild -->
 
     <!-- build:js scripts/config.js -->
@@ -91,6 +95,7 @@
             'superdesk.users',
             'superdesk.planning',
             'superdesk.archive',
+            'superdesk.archive.directives',
             'superdesk.ingest',
             //'superdesk.scratchpad',
             'superdesk.desks',
