+++ client/app/scripts/superdesk/services/all.js
@@ -5,7 +5,6 @@ define([
     './storage',
     './preferencesService',
     './permissionsService',
-    './keyboardManager',
     './entity',
     './server',
     './dragDropService',
@@ -20,7 +19,6 @@ define([
         require('./dragDropService').name,
         require('./server').name,
         require('./entity').name,
-        require('./keyboardManager').name,
         require('./permissionsService').name,
         require('./storage').name,
         require('./preferencesService').name,
