+++ client/app/scripts/superdesk/services/keyboardManager.js
@@ -1,4 +1,4 @@
-define(['angular', 'lodash'], function(angular, _) {
+(function() {
     'use strict';
 
     return angular.module('superdesk.keyboard', [])
@@ -294,4 +294,4 @@ define(['angular', 'lodash'], function(angular, _) {
             }
         };
     }]);
-});
+})();
