+++ client/karma.conf.js
@@ -19,7 +19,7 @@ module.exports = function(config) {
             'app/scripts/bower_components/angular/angular.js',
             {pattern: 'app/**/*.js', included: false},
             {pattern: 'app/scripts/superdesk/**/*[sS]pec.js', included: false},
-            'app/scripts/superdesk-*/**/views/**/*.html',
+            'app/scripts/superdesk-*/**/*.html',
             'app/scripts/superdesk/**/*.html',
             'test-main.js'
         ],
@@ -31,7 +31,7 @@ module.exports = function(config) {
         ],
 
         ngHtml2JsPreprocessor: {
-            prependPrefix: '/base/',
+            stripPrefix: 'app/',
             moduleName: 'templates'
         },
